Post-review cleanup sweep
- Pin `active_periods` array to `int32` in `_validate_initial_regime_active`
  (matches the int32-period invariant established for sibling arrays).
- Hoist `lcm.params` imports in `lcm.utils.containers` to module top; the
  dual lazy imports inside `mapping_leaf` / `sequence_leaf` keep the cycle
  resolved.
- Drop the redundant class-level `@beartype` on `LogSpacedGrid` (the
  `__init__`-level decorator already covers it; matches `LinSpacedGrid`).
- Delete the dead `ValueError` branch in `vmap_1d` (unreachable under the
  package claw's `Literal` check); the test asserts the claw violation
  directly.
- Replace ad-hoc `_persistence.Model = Model` injection in `lcm.__init__`
  with `_bind_forward_refs` helpers in `lcm.persistence` and `lcm.variables`;
  add a regression test that the bindings survive a fresh import.
- Tighten `test_beartype_claw.py` test docstrings to behaviour sentences.
- Reconcile `UserMappingLeaf` "frozen" docstring with `__hash__ = None`.
- Add `UserAge = int | Fraction` alias; sweep `AgeGrid.__init__`,
  `_validate_age_grid`, `_validate_range`, `_validate_values` to use it.
- Trim Claudish prose in the `lcm.__init__` claw block and the
  `_emit_post_loop_diagnostics` / `_raise_first_nan_row` docstrings.

+++ src/lcm/simulation/initial_conditions.py
@@ -435,7 +435,10 @@ def _collect_structural_errors(
         active_mask = jnp.ones(regime_id_arr.size, dtype=bool)
         for regime_name, internal_regime in internal_regimes.items():
             in_regime = regime_id_arr == regime_names_to_ids[regime_name]
-            period_active = jnp.isin(periods, jnp.array(internal_regime.active_periods))
+            period_active = jnp.isin(
+                periods,
+                jnp.array(internal_regime.active_periods, dtype=jnp.int32),
+            )
             active_mask = active_mask & (~in_regime | period_active)
 
         if not jnp.all(active_mask):
